Relocated hint
